Replace data types of `acc_manifest_id`, `agency_id_list_manifest_id`, `agency_id_list_value_manifest_id`, `ascc_manifest_id`, `asccp_manifest_id`, `bcc_manifest_id` `bccp_manifest_id`, `blob_content_manifest_id`, `code_list_manifest_id`, `code_list_value_manifest_id`, `dt_manifest_id`, `dt_sc_manifest_id`, `xbt_manifest_id` with UUID char(36)

+++ score-http/src/main/java/org/oagi/score/repo/component/abie/AbieWriteRepository.java
@@ -45,7 +45,7 @@ public class AbieWriteRepository {
             abieRecord = new AbieRecord();
             abieRecord.setAbieId(UUID.randomUUID().toString());
             abieRecord.setGuid(ScoreGuid.randomGuid());
-            abieRecord.setBasedAccManifestId(ULong.valueOf(abie.getBasedAccManifestId()));
+            abieRecord.setBasedAccManifestId(abie.getBasedAccManifestId());
             abieRecord.setPath(abie.getPath());
             abieRecord.setHashPath(hashPath);
