+++ score-http/src/main/java/org/oagi/score/repo/component/asbie/AsbieWriteRepository.java
@@ -51,7 +51,7 @@ public class AsbieWriteRepository {
             asbieRecord = new AsbieRecord();
             asbieRecord.setAsbieId(UUID.randomUUID().toString());
             asbieRecord.setGuid(ScoreGuid.randomGuid());
-            asbieRecord.setBasedAsccManifestId(ULong.valueOf(asbie.getBasedAsccManifestId()));
+            asbieRecord.setBasedAsccManifestId(asbie.getBasedAsccManifestId());
             asbieRecord.setPath(asbie.getPath());
             asbieRecord.setHashPath(hashPath);
             asbieRecord.setFromAbieId(dslContext.select(ABIE.ABIE_ID)
